spellrite_project/api/permissions.py
# api/permissions.py

from rest_framework import permissions

class IsAdminOrSelf(permissions.BasePermission):
    """
    Custom permission to only allow admins to view any teacher,
    or teachers to view their own information.
    """

    def has_object_permission(self, request, view, obj):
        # Admins have full access
        if request.user.is_staff or request.user.is_superuser:
            return True

        # Teachers can only view their own data
        return obj == request.user
    
class IsOwner(permissions.BasePermission):
    """
    Custom permission to only allow owners of an object to access or edit it.
    """
    def has_object_permission(self, request, view, obj):
        return obj.teacher == request.user