requirements.txt
asgiref==3.8.1
Django==4.2.17
django-environ==0.11.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
mysqlclient==2.2.6
PyJWT==2.10.1
PyMySQL==1.1.1
sqlparse==0.5.3
typing_extensions==4.12.2
